// src/taktflow-api/interfaces/application-handlers.interface.ts
export interface ApplicationHandlers {
  produceEvent:            ProduceEventHandler;
  createTopic:             CreateTopicHandler;
  getTopic:                GetTopicHandler;
  updateTopic:             UpdateTopicHandler;
  deleteTopic:             DeleteTopicHandler;
  createPushConsumer:      CreatePushConsumerHandler;
  createPullConsumer:      CreatePullConsumerHandler;
  createSchedule:          CreateScheduleHandler;
  createApiKey:            CreateApiKeyHandler;
  listApiKeys:             ListApiKeysHandler;
  getApiKey:               GetApiKeyHandler;
  deleteApiKey:            DeleteApiKeyHandler;
  registerTenant:          RegisterTenantHandler;
  listEvents:              ListEventsHandler;
  getEventDetail:          GetEventDetailHandler;
  listTopics:              ListTopicsHandler;
  listConsumers:           ListConsumersHandler;
  listSchedules:           ListSchedulesHandler;
  getDashboardMetrics:     GetDashboardMetricsHandler;
  login:                   LoginHandler;
  refresh:                 RefreshTokenHandler;
  logout:                  LogoutHandler;
  createUser:              CreateUserHandler;
  getCurrentUser:          GetCurrentUserHandler;
  updateUser:              UpdateUserHandler;
  changePassword:          ChangePasswordHandler;
  consumeEvents:           ConsumeEventsHandler;
  getTenant:               GetTenantHandler;
  updateTenant:            UpdateTenantHandler;
  getUsage:                GetUsageHandler;
  listDeadLetterEvents:    ListDeadLetterEventsHandler;
  replayDeadLetterEvent:   ReplayDeadLetterEventHandler;
  getConsumerHealth:       GetConsumerHealthHandler;
  getConsumer:             GetConsumerHandler;
  updateConsumer:          UpdateConsumerHandler;
  deleteConsumer:          DeleteConsumerHandler;
  pauseConsumer:           PauseConsumerHandler;
  resumeConsumer:          ResumeConsumerHandler;
  acknowledgeEvent:        AcknowledgeEventHandler;
}